Add RAS_AGENT Service Group
In ACPI systems, The RAS_AGENT works as a bridge
between the RISCV RERI compliant error sources in
the system and the APEI tables. One of its job is
the populate CPER records when errors occur.

The RAS_AGENT has complete knowledge of the error
sources present in the system and errors they can
generate. The services in the RAS_AGENT service group
respond to
- Enumeration of various error sources present in the system
- Get the error source descriptor for an error source

These services from RAS_AGENT can be used to populate
the HEST table when ACPI tables are being generated.

For non-ACPI systems also, the error sources can be
enumerated. The format for the error source descriptor
is left to the implementation.

Signed-off-by: Himanshu Chauhan <hchauhan@ventanamicro.com>
Signed-off-by: Rahul Pathak <rpathak@ventanamicro.com>