logging: Add Logging service group
This service group defines services used by an application processor to set
the log data defined by the UEFI Platform Initialization specification or
other similar implementations. This service group provides a reliable way
to track the progress of the system for diagnostic purposes. The data
provided by the services can then be relayed to other consumers in the
system such as the BMC, LCD display, etc.

Signed-off-by: Subrahmanya Lingappa <subrahmanya.lingappa@qti.qualcomm.com>